[refactor] CertificationNumberVC와 CertificationNumberView를 분리
